synchronize works for both cpu and gpu
