+++ docs/src/python/devices_and_streams.rst
@@ -16,3 +16,4 @@ Devices and Streams
    new_stream
    set_default_stream
    stream
+   synchronize
