+++ docs/src/python/metal.rst
@@ -7,7 +7,6 @@ Metal
   :toctree: _autosummary
 
   is_available
-  synchronize
   get_active_memory
   get_peak_memory
   get_cache_memory
