+++ mlx/backend/metal/metal_impl.h
@@ -2,6 +2,7 @@
 
 #pragma once
 
+#include <future>
 #include <memory>
 
 #include "mlx/array.h"
@@ -14,5 +15,9 @@ void new_stream(Stream stream);
 std::unique_ptr<void, std::function<void(void*)>> new_scoped_memory_pool();
 
 std::function<void()> make_task(array arr, bool signal);
+
+std::function<void()> make_synchronize_task(
+    Stream s,
+    std::shared_ptr<std::promise<void>> p);
 
 } // namespace mlx::core::metal
